Refactor src/crud.py into a package for better cohesion
